requirements.txt
Flask==2.3.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
python-dotenv==1.0.1
grpcio
